interface/requirements.txt
streamlit==1.32.0
pydantic==2.6.1
requests==2.31.0
python-dotenv==1.0.1
openai==1.12.0
httpx==0.27.0 